feat: able to collapse all group in example data
