+++ svelte/src/lib/event-handler.ts
@@ -1,23 +1,25 @@
-export function dragStartedNode(event: any) {
-	// if (!event.active) simulation.alpha(0.7).restart();
+export function dragStartedNode(event: any, simulation: any) {
+	if (!event.active) simulation.alpha(0.05).restart();
 	event.subject.fx = event.subject.x;
 	event.subject.fy = event.subject.y;
 }
 
 // Update the subject (dragged node) position during drag.
-export function draggedNode(event: any) {
+export function draggedNode(event: any, simulation: any) {
+	simulation.alpha(0.05).restart();
+	console.log('test');
 	event.subject.fx = event.x;
 	event.subject.fy = event.y;
 }
 
-export function dragEndedNode(event: any) {
-	// if (!event.active) simulation.alpha(config.isForceSimulationEnabled ? 0.7 : 0).restart();
+export function dragEndedNode(event: any, simulation: any) {
+	if (!event.active) simulation.alpha(0.05).restart();
 	event.subject.fx = null;
 	event.subject.fy = null;
 }
 
-export function dragStartedGroup(event: any) {
-	// if (!event.active) simulation.alpha(0.7).restart();
+export function dragStartedGroup(event: any, simulation: any) {
+	if (!event.active) simulation.alpha(0.05).restart();
 	event.subject.fx = event.subject.x;
 	event.subject.fy = event.subject.y;
 	event.subject.members.forEach((member: any) => {
@@ -26,7 +28,9 @@ export function dragStartedGroup(event: any) {
 	});
 }
 
-export function draggedGroup(event: any) {
+export function draggedGroup(event: any, simulation: any) {
+	simulation.alpha(0.05).restart();
+
 	// console.log(event.subject.y, event.y);
 	const differenceX = event.x - event.subject.fx;
 	const differenceY = event.y - event.subject.fy;
@@ -39,8 +43,8 @@ export function draggedGroup(event: any) {
 		member.fy += differenceY;
 	});
 }
-export function dragEndedGroup(event: any) {
-	// if (!event.active) simulation.alpha(config.isForceSimulationEnabled ? 0.7 : 0).restart();
+export function dragEndedGroup(event: any, simulation: any) {
+	if (!event.active) simulation.alpha(0.05).restart();
 	event.subject.fx = null;
 	event.subject.fy = null;
 	event.subject.members.forEach((member: any) => {
