+++ svelte/src/routes/components/side-panel.svelte
